Conflicts:
	admin-dev/themes/default/css/admin-theme.css

+++ admin-dev/themes/default/template/controllers/categories/helpers/form/form.tpl
@@ -34,7 +34,7 @@
 					{if $shared_category}
 						<p class="alert alert-warning">{l s='If you delete this picture, it will be deleted in all of your shared shops!'}</p>
 					{/if}
-					<a class="btn btn-default" href="{$current}&{$identifier}={$form_id}&token={$token}&{if $shared_category}forcedeleteImage=1{else}deleteImage=1{/if}">
+					<a class="btn btn-default" href="{$current}&{$identifier}={$form_id}&token={$token}&{if $shared_category}forcedeleteImage=1{else}deleteImage=1{/if}" onclick="return confirm('{l s='If you delete this picture, it will be deleted in all of your shared shops!'}\n\n{l s='Do you want to continue?'}')">
 						<i class="icon-trash"></i> {l s='Delete'}
 					</a>
 				</div>
